organized pages and removed audio file
