+++ app/.gitignore
@@ -35,9 +35,3 @@ lib/generated_plugin_registrant.dart
 
 # Exceptions to above rules.
 !/packages/flutter_tools/test/data/dart_dependencies_test/**/.packages
-
-/app/assets/songExample.mp3
-/app/audio/songExample.mp3
-
-
-
