+++ app/lib/about_page/TitleSection.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class TitleSection extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Container(
+        margin: EdgeInsets.only(top: 15, bottom: 15),
+        child: Column(
+          children: <Widget>[
+            Text(
+              "Katy Trail History",
+              style: TextStyle(
+                color: Colors.black,
+                fontWeight: FontWeight.bold,
+                fontSize: 30,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
