+++ app/pubspec.yaml
@@ -39,7 +39,7 @@ flutter:
    
   assets:     
     - assets/images/katy_trail.jpg
-    - assets/songExample.mp3
+    #- assets/songExample.mp3
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
